Improve code organization for coverage viz pipeline step.
Removed util/coverage and moved functions into PipelineStepGenerateCoverageViz,
since those functions are specific to that one pipeline step.

Reorganized functions in PipelineStepGenerateCoverageViz so that the "main"
functions are at the top of the file.

Consolidated tests into one file.
